// src/api/models/collection.ts
/* tslint:disable */
/* eslint-disable */
/**
 * IMX Market Trader
 * IMX Market Trader backend API
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
/**
 * 
 * @export
 * @interface Collection
 */
export interface Collection {
    /**
     * 
     * @type {string}
     * @memberof Collection
     */
    _id: string;
    /**
     * 
     * @type {string}
     * @memberof Collection
     */
    collectionAddress: string;
}