// src/api/models/delete-result.ts
/* tslint:disable */
/* eslint-disable */
/**
 * IMX Market Trader
 * IMX Market Trader backend API
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
/**
 * 
 * @export
 * @interface DeleteResult
 */
export interface DeleteResult {
    /**
     * 
     * @type {boolean}
     * @memberof DeleteResult
     */
    acknowledged: boolean;
    /**
     * 
     * @type {number}
     * @memberof DeleteResult
     */
    deletedCount: number;
}